update the workspace command line argument to store true, update all the types to float

+++ src/activation.cpp
@@ -1,6 +1,6 @@
 #include "activation.h"
 
-int apply_relu(double *input, int size) {
+int apply_relu(float *input, int size) {
     for (int i = 0; i < size; i++) {
         if (input[i] < 0) {
             input[i] = 0;
